pyproject.toml
[project]
name = "langchain-learn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "chromadb>=1.3.5",
    "langchain>=1.1.0",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.0",
    "langgraph-checkpoint-postgres>=3.0.2",
    "psycopg>=3.3.2",
    "psycopg-binary>=3.3.2",
    "pypdf>=6.4.1",
]

[[tool.uv.index]]
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
default = true
